+++ packages/backend-deployer/src/cdk_error_mapper.ts
@@ -44,7 +44,10 @@ export class CdkErrorMapper {
 
     if (error.message.includes('does not support module.register()')) {
       let resolutionMessage;
-      if (deploymentType === 'branch') {
+      if (deploymentType === 'standalone') {
+        resolutionMessage =
+          'Upgrade the node version in your CI/CD environment to `^18.19.0`, `^20.6.0`, or `>=22`.';
+      } else if (deploymentType === 'branch') {
         resolutionMessage =
           'Upgrade the node version in your CI/CD environment. ' +
           'If you are using Amplify Hosting for your backend builds, you can add `nvm install 18.x` or `nvm install 20.x` in your `amplify.yml` before the `pipeline-deploy` command';