+++ packages/cli/src/commands/generate/forms/generate_forms_command.ts
@@ -26,9 +26,10 @@ type GenerateFormsCommandOptionsCamelCase = {
 /**
  * Command that generates UI forms.
  */
-export class GenerateFormsCommand
-  implements CommandModule<object, GenerateFormsCommandOptions>
-{
+export class GenerateFormsCommand implements CommandModule<
+  object,
+  GenerateFormsCommandOptions
+> {
   /**
    * @inheritDoc
    */