+++ packages/cli/src/commands/sandbox/sandbox-secret/sandbox_secret_set_command.ts
@@ -11,9 +11,10 @@ import { AmplifyUserError } from '@aws-amplify/platform-core';
 /**
  * Command to set sandbox secret.
  */
-export class SandboxSecretSetCommand
-  implements CommandModule<object, SecretSetCommandOptionsKebabCase>
-{
+export class SandboxSecretSetCommand implements CommandModule<
+  object,
+  SecretSetCommandOptionsKebabCase
+> {
   /**
    * @inheritDoc
    */