+++ packages/integration-tests/src/test-projects/data-storage-auth-with-triggers-ts/amplify/test_factories.ts
@@ -1,5 +1,5 @@
 import { data } from './data/resource.js';
-import { defaultNodeFunc, node16Func } from './function.js';
+import { defaultNodeFunc } from './function.js';
 import { storage } from './storage/resource.js';
 import { auth } from './auth/resource.js';
 
@@ -8,5 +8,4 @@ export const dataStorageAuthWithTriggers = {
   storage,
   defaultNodeFunc,
   data,
-  node16Func,
 };