+++ packages/integration-tests/src/test-projects/data-storage-auth-with-triggers-ts/hotswap-update-files/function.ts
@@ -15,19 +15,6 @@ export const defaultNodeFunc = defineFunction({
   timeoutSeconds: 5,
 });
 
-export const node16Func = defineFunction({
-  name: 'node16Function',
-  entry: './func-src/handler_node16.ts',
-  environment: {
-    TEST_SECRET: secret('amazonSecret'),
-    TEST_SHARED_SECRET: secret(
-      process.env[amplifySharedSecretNameKey] as string,
-    ),
-  },
-  timeoutSeconds: 5,
-  runtime: 16,
-});
-
 export const onDelete = defineFunction({
   name: 'onDelete',
   entry: './func-src/handler.ts',