+++ packages/backend-auth/CHANGELOG.md
@@ -1,5 +1,33 @@
 # @aws-amplify/backend-auth
 
+## 1.9.3
+
+### Patch Changes
+
+- 1d109e9: remove aws-lambda package, @types/aws-lambda is sufficent
+
+## 1.9.2
+
+### Patch Changes
+
+- 4c5dd61: Update default function version to 22 and add Node 24 as runtime option, additionally update all functions that use Node 20 to Node 22
+- Updated dependencies [4c5dd61]
+  - @aws-amplify/auth-construct@1.11.1
+
+## 1.9.1
+
+### Patch Changes
+
+- 7d0ba5e: chore: upgrade CDK dependencies
+- 4603f7a: bump aws-cdk-lib version to ^2.234.1 across all packages
+- Updated dependencies [7d0ba5e]
+- Updated dependencies [299c804]
+- Updated dependencies [4603f7a]
+  - @aws-amplify/backend-output-storage@1.3.3
+  - @aws-amplify/auth-construct@1.11.0
+  - @aws-amplify/plugin-types@1.11.2
+  - @aws-amplify/backend-output-schemas@1.8.0
+
 ## 1.9.0
 
 ### Minor Changes