+++ packages/cli/CHANGELOG.md
@@ -1,5 +1,20 @@
 # @aws-amplify/backend-cli
 
+## 1.8.2
+
+### Patch Changes
+
+- 7d0ba5e: chore: upgrade CDK dependencies
+- 4603f7a: bump aws-cdk-lib version to ^2.234.1 across all packages
+- Updated dependencies [7d0ba5e]
+- Updated dependencies [299c804]
+- Updated dependencies [4603f7a]
+  - @aws-amplify/backend-deployer@2.1.5
+  - @aws-amplify/platform-core@1.10.4
+  - @aws-amplify/plugin-types@1.11.2
+  - @aws-amplify/backend-output-schemas@1.8.0
+  - @aws-amplify/client-config@1.10.0
+
 ## 1.8.1
 
 ### Patch Changes