+++ packages/platform-core/CHANGELOG.md
@@ -1,5 +1,27 @@
 # @aws-amplify/platform-core
 
+## 1.11.0
+
+### Minor Changes
+
+- 67e8773: Add standalone deployment type for deploying Gen2 backends without Amplify Hosting
+
+### Patch Changes
+
+- Updated dependencies [0ee9189]
+- Updated dependencies [67e8773]
+  - @aws-amplify/plugin-types@1.12.0
+
+## 1.10.4
+
+### Patch Changes
+
+- 7d0ba5e: chore: upgrade CDK dependencies
+- 4603f7a: bump aws-cdk-lib version to ^2.234.1 across all packages
+- Updated dependencies [7d0ba5e]
+- Updated dependencies [4603f7a]
+  - @aws-amplify/plugin-types@1.11.2
+
 ## 1.10.3
 
 ### Patch Changes